fix: order and order_items

+++ app/Http/Controllers/Api/Order/OrderItemController.php
@@ -16,7 +16,8 @@ class OrderItemController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-    public function orderItem(OrderItem $orderItem){
+    public function orderItem(OrderItem $orderItem)
+    {
         try {
             $this->authorize('getOrder', $orderItem->order);
 
@@ -24,41 +25,46 @@ class OrderItemController extends Controller
                 'status' => 'success',
                 'data' => $orderItem->makeHidden(['order', 'product']),
             ], 201);
-
         } catch (AuthorizationException $e) {
             return response()->json([
                 'status' => 'error',
                 'message' => 'Vous n\'êtes pas autorisé à faire cette action.',
             ], 403);
-        }        
-        catch (Exception $e) {
+        } catch (Exception $e) {
             return response()->json($e);
         }
     }
 
 
-    public function delete(OrderItem $orderItem){
-
+    public function delete(OrderItem $orderItem)
+    {
         try {
+            $order = $orderItem->order;
             $this->authorize('delete', $orderItem->order);
 
-            // solf delete
+            if(!$orderItem || !$order){
+                return response()->json([
+                    'message' => 'Commande non trouvée.'
+                ], 400);
+            }
+
+            if ($order->order_items->count() == 1){
+                $order->delete();
+            }
+
             $orderItem->delete();
 
             return response()->json([
                 'status' => 'success',
                 'message' => 'Suppression effectuée avec succès.'
             ], 200);
-
-        } catch(AuthorizationException $e){
+        } catch (AuthorizationException $e) {
             return response()->json([
                 'status' => 'error',
                 'message' => 'Vous n\'êtes pas autorisé à faire cette action.',
             ], 403);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             return response()->json($e);
         }
     }
-
-
 }
